requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
pytest==8.4.0
requests==2.32.4
